Collect remaining nodes found in PG15's regression test suite
Since the parser is on PG13, there are some syntax errors on queries,
namely WITH (...) SEARCH and CYCLE, though these shouldn't appear as
nodes that sqlreduce will later trip on, unless specifically directed to
descend into these attributes.

We also move all the test-only no-op nodes to the bottom of rules_yaml.